move isExtensionAdded to common file
